+++ src/desktop/src/ui/global/Polling.js
@@ -54,8 +54,6 @@ class Polling extends React.PureComponent {
         /** @ignore */
         setPollFor: PropTypes.func.isRequired,
         /** @ignore */
-        marketData: PropTypes.object.isRequired,
-        /** @ignore */
         fetchMarketData: PropTypes.func.isRequired,
         /** @ignore */
         fetchPrice: PropTypes.func.isRequired,
@@ -109,17 +107,6 @@ class Polling extends React.PureComponent {
 
         this.onPollTick = this.fetch.bind(this);
         this.interval = setInterval(this.onPollTick, 8000);
-    }
-
-    componentDidUpdate(prevProps) {
-        const { marketData, isPollingMarketData } = this.props;
-
-        /**
-         * Send updated marketData to Tray application
-         */
-        if (prevProps.isPollingMarketData && !isPollingMarketData) {
-            Electron.storeUpdate(JSON.stringify({ marketData }));
-        }
     }
 
     componentWillUnmount() {
@@ -258,7 +245,6 @@ const mapStateToProps = (state) => ({
     isRetryingFailedTransaction: state.ui.isRetryingFailedTransaction,
     failedBundleHashes: getFailedBundleHashes(state),
     password: state.wallet.password,
-    marketData: state.marketData,
 });
 
 const mapDispatchToProps = {