+++ src/mobile/src/ui/components/Poll.js
@@ -129,25 +129,23 @@ export class Poll extends Component {
     };
 
     shouldSkipCycle() {
-        const props = this.props;
-
         const isAlreadyDoingSomeHeavyLifting =
-            props.isSyncing ||
-            props.isSendingTransfer ||
-            props.isGeneratingReceiveAddress ||
-            props.isFetchingAccountInfo || // In case the app is already fetching latest account info, stop polling because the market related data is already fetched on login
-            props.addingAdditionalAccount ||
-            props.isTransitioning ||
-            props.isPromotingTransaction ||
-            props.isRetryingFailedTransaction;
+            this.props.isSyncing ||
+            this.props.isSendingTransfer ||
+            this.props.isGeneratingReceiveAddress ||
+            this.props.isFetchingAccountInfo || // In case the app is already fetching latest account info, stop polling because the market related data is already fetched on login
+            this.props.addingAdditionalAccount ||
+            this.props.isTransitioning ||
+            this.props.isPromotingTransaction ||
+            this.props.isRetryingFailedTransaction;
 
         const isAlreadyPollingSomething =
-            props.isPollingPrice ||
-            props.isPollingNodeList ||
-            props.isPollingChartData ||
-            props.isPollingMarketData ||
-            props.isPollingAccountInfo ||
-            props.isAutoPromoting;
+            this.props.isPollingPrice ||
+            this.props.isPollingNodeList ||
+            this.props.isPollingChartData ||
+            this.props.isPollingMarketData ||
+            this.props.isPollingAccountInfo ||
+            this.props.isAutoPromoting;
 
         return isAlreadyDoingSomeHeavyLifting || isAlreadyPollingSomething;
     }