+++ src/shared/libs/iota/extendedApi.js
@@ -463,7 +463,13 @@ const checkAttachToTangleAsync = (node) => {
             'X-IOTA-API-Version': IRI_API_VERSION,
         }),
     })
-        .then((res) => res.json())
+        .then((response) => {
+            if (response.ok) {
+                return response.json();
+            }
+
+            throw response;
+        })
         .catch(() => {
             // return a fake normal IRI response when attachToTangle is not available
             return { error: Errors.ATTACH_TO_TANGLE_UNAVAILABLE };