+++ src/shared/storage/index.js
@@ -97,7 +97,7 @@ class Account {
         if (some(orderedAccounts, (account, index) => index !== account.index)) {
             realm.write(() => {
                 each(orderedAccounts, (account, index) => {
-                    realm.create('Account', assign({}, account, { index }), true);
+                    realm.create('Account', assign({}, account, { index }), 'modified');
                 });
             });
         }
@@ -161,7 +161,7 @@ class Account {
                       ),
             });
 
-            realm.create('Account', updatedData, true);
+            realm.create('Account', updatedData, 'modified');
         });
     }
 
@@ -188,7 +188,7 @@ class Account {
                         assign({}, account, {
                             index: account.index > deletedAccountIndex ? account.index - 1 : account.index,
                         }),
-                        true,
+                        'modified',
                     );
                 });
             }
@@ -295,7 +295,7 @@ class Node {
                 each(nodes, (node) => {
                     // If it's an existing node, just update properties.
                     if (includes(existingUrls, node.url)) {
-                        realm.create('Node', node, true);
+                        realm.create('Node', node, 'modified');
                     } else {
                         realm.create('Node', node);
                     }
@@ -388,6 +388,18 @@ class Wallet {
     static updateRemotePowSetting(payload) {
         realm.write(() => {
             Wallet.latestSettings.remotePoW = payload;
+        });
+    }
+
+    /**
+     * Updates proof of work node setting.
+     *
+     * @method updatePowNodeSetting
+     * @param {string} payload
+     */
+    static updatePowNodeSetting(payload) {
+        realm.write(() => {
+            Wallet.latestSettings.powNode = payload;
         });
     }
 
@@ -769,7 +781,7 @@ class Wallet {
             realm.write(() =>
                 realm.create('Wallet', {
                     version: Wallet.version,
-                    settings: { notifications: {}, quorum: {} },
+                    settings: { notifications: {}, quorum: {}, powNode: '' },
                     accountInfoDuringSetup: { meta: {} },
                 }),
             );
@@ -790,7 +802,7 @@ class Wallet {
                     version: Wallet.version,
                     ...merge({}, Wallet.latestData, data),
                 },
-                true,
+                'modified',
             ),
         );
     }